+++ src/appleseed-max-impl/appleseed-max2016-impl.vcxproj
@@ -243,12 +243,12 @@ copy /Y "$(SolutionDir)..\..\appleseed\sandbox\shaders\appleseed\*.oso" "$(Solut
     <ClCompile Include="appleseedinteractive\interactivetilecallback.cpp" />
     <ClCompile Include="appleseedmetalmtl\appleseedmetalmtl.cpp" />
     <ClCompile Include="appleseedobjpropsmod\appleseedobjpropsmod.cpp" />
-    <ClCompile Include="appleseedosltexture\osltexture.cpp" />
-    <ClCompile Include="appleseedosltexture\oslclassdesc.cpp" />
-    <ClCompile Include="appleseedosltexture\oslmaterial.cpp" />
-    <ClCompile Include="appleseedosltexture\oslshadermetadata.cpp" />
-    <ClCompile Include="appleseedosltexture\oslshaderregistry.cpp" />
-    <ClCompile Include="appleseedosltexture\oslparamdlg.cpp" />
+    <ClCompile Include="appleseedoslplugin\osltexture.cpp" />
+    <ClCompile Include="appleseedoslplugin\oslclassdesc.cpp" />
+    <ClCompile Include="appleseedoslplugin\oslmaterial.cpp" />
+    <ClCompile Include="appleseedoslplugin\oslshadermetadata.cpp" />
+    <ClCompile Include="appleseedoslplugin\oslshaderregistry.cpp" />
+    <ClCompile Include="appleseedoslplugin\oslparamdlg.cpp" />
     <ClCompile Include="appleseedplasticmtl\appleseedplasticmtl.cpp" />
     <ClCompile Include="appleseedrenderer\dialoglogtarget.cpp" />
     <ClCompile Include="iappleseedmtl.cpp" />
@@ -286,13 +286,13 @@ copy /Y "$(SolutionDir)..\..\appleseed\sandbox\shaders\appleseed\*.oso" "$(Solut
     <ClInclude Include="appleseedmetalmtl\resource.h" />
     <ClInclude Include="appleseedobjpropsmod\appleseedobjpropsmod.h" />
     <ClInclude Include="appleseedobjpropsmod\resource.h" />
-    <ClInclude Include="appleseedosltexture\osltexture.h" />
-    <ClInclude Include="appleseedosltexture\oslclassdesc.h" />
-    <ClInclude Include="appleseedosltexture\oslmaterial.h" />
-    <ClInclude Include="appleseedosltexture\oslshadermetadata.h" />
-    <ClInclude Include="appleseedosltexture\oslshaderregistry.h" />
-    <ClInclude Include="appleseedosltexture\oslparamdlg.h" />
-    <ClInclude Include="appleseedosltexture\templategenerator.h" />
+    <ClInclude Include="appleseedoslplugin\osltexture.h" />
+    <ClInclude Include="appleseedoslplugin\oslclassdesc.h" />
+    <ClInclude Include="appleseedoslplugin\oslmaterial.h" />
+    <ClInclude Include="appleseedoslplugin\oslshadermetadata.h" />
+    <ClInclude Include="appleseedoslplugin\oslshaderregistry.h" />
+    <ClInclude Include="appleseedoslplugin\oslparamdlg.h" />
+    <ClInclude Include="appleseedoslplugin\templategenerator.h" />
     <ClInclude Include="appleseedplasticmtl\appleseedplasticmtl.h" />
     <ClInclude Include="appleseedplasticmtl\datachunks.h" />
     <ClInclude Include="appleseedplasticmtl\resource.h" />
