+++ src/appleseed-max-impl/appleseedoslplugin/oslclassdesc.h
