+++ src/appleseed-max-impl/appleseedoslplugin/oslshadermetadata.cpp
