+++ src/appleseed-max-impl/appleseedoslplugin/oslshadermetadata.h
