+++ src/appleseed-max-impl/appleseedoslplugin/templategenerator.h
